simplify data_rw returns and fix lzw file_sz issue

+++ aes/aes.h
@@ -6,7 +6,7 @@
 /*   By: herrfalco <marvin@42.fr>                   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/04 23:12:22 by herrfalco         #+#    #+#             */
-/*   Updated: 2022/06/25 13:10:37 by fcadet           ###   ########.fr       */
+/*   Updated: 2022/06/25 18:18:07 by fcadet           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,6 +27,6 @@ int		rand_key(uint8_t *buff, size_t size);
 void	round_keys(uint8_t *key, uint32_t *rkeys);
 
 void	aes_data_enc(uint8_t *data, uint64_t size, uint32_t *r_keys);
-void	aes_fd_enc(int fd_dst, int fd_src, uint8_t *key);
+void	aes_fd_enc(int64_t fd_dst, int64_t fd_src, uint8_t *key);
 
 #endif // AES_H
